% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathcat.R
\name{file_path_sans_all_ext}
\alias{file_path_sans_all_ext}
\title{Removes all extensions from the file. Based on the tools::file_path_sans_ext}
\usage{
file_path_sans_all_ext(path)
}
\arguments{
\item{path}{Path from which the extensions will be removed}
}
\value{
Character string with the path without all extensions from the \code{basename(path)} removed.
}
\description{
Removes all extensions from the file. Based on the tools::file_path_sans_ext
}
