% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathcat.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{Checks whether a given path is absolute, or not}
\usage{
is_absolute_path(path)
}
\arguments{
\item{path}{Path to assess.}
}
\value{
Boolean
}
\description{
Checks whether a given path is absolute, or not
}
