% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathcat.R
\name{make.path.relative}
\alias{make.path.relative}
\title{Converts one absolute path into relative}
\usage{
make.path.relative(base.path, target.path, fsep = .Platform$file.sep)
}
\arguments{
\item{base.path}{character string with starting directory.}

\item{fsep}{Custom path separator. Defaults to \code{.Platform$file.sep}.}

\item{target}{character string with target directory.}
}
\value{
Character string that represents a relative path that leads to \code{target.path}
from \code{base.path}.
}
\description{
The algorithm produces a relative path, that addresses \code{target.path} from
\code{base.path} using folder names and special name \code{..} for parent folder.
}
