% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathcat.R
\docType{package}
\name{pathcat}
\alias{pathcat}
\alias{pathcat-package}
\title{Provides path.cat and make.path.relative functions, that enhance \code{base::system.path()}.}
\description{
\code{path.cat} concatenates paths in a way that mimmics OS behaviour when
inputing subsequent path components into separate \code{setwd} commands.
}
\details{
All functions allow for mixing '/' with '\' on each path, treating both characters
 equally (so you can't have a filename with backslash even if you are on Linux)
}

