% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathcat.R
\name{split.absolute.prefix}
\alias{split.absolute.prefix}
\title{Splits the absolute path into the absolute part (e.g. '/' or 'H:\\') and the rest.
If not absolute path, return undefined}
\usage{
\method{split}{absolute.prefix}(abs.path)
}
\description{
Splits the absolute path into the absolute part (e.g. '/' or 'H:\\') and the rest.
If not absolute path, return undefined
}
