% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathcat.R
\name{which.paths.are.absolute}
\alias{which.paths.are.absolute}
\title{Returns indices of paths that are absolute}
\usage{
which.paths.are.absolute(paths)
}
\description{
Returns indices of paths that are absolute
}
